feat: kafka state store 추가
